backend/requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
httpx
python-multipart 